show confirm dialog when delete a snippet

+++ components/SnippetItem.jsx
@@ -5,6 +5,7 @@ import CardActions from 'material-ui/lib/card/card-actions';
 import CardHeader from 'material-ui/lib/card/card-header';
 import FlatButton from 'material-ui/lib/flat-button';
 import CardText from 'material-ui/lib/card/card-text';
+import Dialog from 'material-ui/lib/dialog';
 import { destroySnippet } from '../actions/SnippetActions';
 
 const langIcons = {
@@ -20,16 +21,23 @@ class SnippetItem extends Component {
   constructor (props) {
     super(props);
     this.state = {
-      open: false,
+      openDialog: false
     };
 
     this.handleClickDeleteButton = this.handleClickDeleteButton.bind(this);
+    this.handleCloseDialog = this.handleCloseDialog.bind(this);
+  }
+
+  handleCloseDialog (id, confirm) {
+    this.setState({
+      openDialog: false
+    });
+    if (confirm) this.props.dispatch(destroySnippet(id));
   }
 
   handleClickDeleteButton (id) {
-    this.props.dispatch(destroySnippet(id));
     this.setState({
-      open: true,
+      openDialog: true
     });
   }
 
@@ -43,20 +51,42 @@ class SnippetItem extends Component {
     );
     const snippetAction = (
       <CardActions>
-        <FlatButton label="Delete" primary={true} onClick={this.handleClickDeleteButton.bind(null, metaData._id)} />
+        <FlatButton label="Delete" primary={true} onClick={this.handleClickDeleteButton} />
       </CardActions>
     )
+    const dialogActions = [
+      <FlatButton
+        label           = "Cancel"
+        secondary       = {true}
+        keyboardFocused = {true}
+        onTouchTap      = {this.handleCloseDialog.bind(null, metaData._id, false)}
+      />,
+      <FlatButton
+        label      = "Submit"
+        primary    = {true}
+        onTouchTap = {this.handleCloseDialog.bind(null, metaData._id, true)}
+      />,
+    ];
 
     return (
-      <Card>
-        <CardHeader
-          title    = {metaData.title}
-          subtitle = {`@ ${metaData.time.toLocaleString()}`}
-          avatar   = {langIcons[metaData.lang]}
-        />
-        {expand ? snippetDetail : null}
-        {expand ? snippetAction : null}
-      </Card>
+      <div>
+        <Card>
+          <CardHeader
+            title    = {metaData.title}
+            subtitle = {`@ ${metaData.time.toLocaleString()}`}
+            avatar   = {langIcons[metaData.lang]}
+          />
+          {expand ? snippetDetail : null}
+          {expand ? snippetAction : null}
+        </Card>
+        <Dialog
+          title   = "Dialog With Actions"
+          open    = {this.state.openDialog}
+          actions = {dialogActions}
+        >
+          The actions in this window were passed in as an array of React objects.
+        </Dialog>
+      </div>
     );
   }
 }
